Fix race condition in SocketBuffer + isForceSending
If a number of threads were trying to write at the end of the buffer
there was a nasty race condition that could leave "holes" in the buffer
due to the ordering of decrements and increments.

This change never decrements the writeOffset, and simply tracks the
maximum position written to via another AtomicInt.

This also protects against trying to force send when we're already force
sending by using an AtomicBoolean